Split tram_transit.rs into subdirectory module (960 → 3 files)
Split the monolithic tram_transit.rs (960 lines) into a tram_transit/
subdirectory with three focused sub-modules:
- state.rs (321 lines): data types, constants, and TramTransitState API
- systems.rs (412 lines): ECS systems, Saveable impls, and plugin
- tests.rs (224 lines): all unit tests

All public items are re-exported from mod.rs for backward compatibility.

Part of #1389
